Split LocatorTestModal into three components
